requirements.txt
# Core deep learning frameworks
torch>=1.10.0
torchvision>=0.11.0
timm>=0.4.12
transformers>=4.15.0
diffusers>=0.10.0
accelerate>=0.6.0
safetensors>=0.2.0

# Image processing
pillow>=8.3.0
opencv-python>=4.5.3
segment-anything>=1.0
matplotlib>=3.4.0

# Data processing
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.62.0

# File handling
requests>=2.26.0
rarfile>=4.0
zipfile38>=0.0.3
gdown>=4.4.0

# Configuration
pyyaml>=6.0
omegaconf>=2.1.1

# Visualization and logging
wandb>=0.12.0

# Model optimization
xformers>=0.0.16
bitsandbytes>=0.35.0

# Compatibility
importlib-metadata>=4.8.1